Project/Part-3/Consumer/requirements.txt
requests
certifi
confluent-kafka[avro,json,protobuf]>=1.4.2
pandas
psycopg2-binary
